refactor: UI 내부 패키지 구조 수정
