+++ app/src/androidTest/java/woowacourse/movie/ui/seat/BookingSeatActivityTest.kt
@@ -17,6 +17,7 @@ import org.junit.Test
 import woowacourse.movie.R
 import woowacourse.movie.domain.model.Headcount
 import woowacourse.movie.fixture.fakeContext
+import woowacourse.movie.ui.seat.view.BookingSeatActivity
 import java.time.LocalDateTime
 
 class BookingSeatActivityTest {
