+++ app/src/main/AndroidManifest.xml
@@ -16,13 +16,13 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".ui.seat.BookingSeatActivity"
+            android:name=".ui.seat.view.BookingSeatActivity"
             android:exported="false" />
         <activity
             android:name=".ui.booking.view.BookingActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.complete.BookingCompleteActivity"
+            android:name=".ui.complete.view.BookingCompleteActivity"
             android:exported="false" />
         <activity
             android:name=".ui.movielist.view.MovieListActivity"
