+++ app/src/main/java/woowacourse/movie/ui/complete/presenter/BookingCompletePresenter.kt
@@ -1,8 +1,9 @@
-package woowacourse.movie.ui.complete
+package woowacourse.movie.ui.complete.presenter
 
 import woowacourse.movie.domain.model.BookedTicket
 import woowacourse.movie.domain.model.Headcount
 import woowacourse.movie.domain.model.Seats
+import woowacourse.movie.ui.complete.contract.BookingCompleteContract
 import java.time.LocalDateTime
 
 class BookingCompletePresenter(
