+++ app/src/main/res/layout/activity_booking_complete.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.complete.BookingCompleteActivity">
+    tools:context=".ui.complete.view.BookingCompleteActivity">
 
     <LinearLayout
         android:id="@+id/linearLayout"
@@ -38,48 +38,44 @@
         android:textSize="28sp"
         android:textStyle="bold"
         app:layout_constraintTop_toBottomOf="@id/linearLayout"
-        tools:text="해리 포터와 마법사의 돌"
-        />
+        tools:text="해리 포터와 마법사의 돌" />
 
     <TextView
         android:id="@+id/tv_release_date"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
         android:gravity="center"
         android:textColor="@color/black"
         android:textSize="20sp"
         android:textStyle="bold"
-        android:layout_marginTop="20dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/tv_title"
-        tools:text="2025.4.15 11:00"
-        />
+        tools:text="2025.4.15 11:00" />
 
     <TextView
         android:id="@+id/tv_headcount"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginTop="50dp"
         android:textColor="@color/black"
         android:textSize="28sp"
-        android:layout_marginTop="50dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/tv_release_date"
-        tools:text="일반 2명"
-        />
+        tools:text="일반 2명" />
 
     <TextView
         android:id="@+id/tv_price"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginTop="26dp"
         android:textColor="@color/black"
         android:textSize="28sp"
-        android:layout_marginTop="26dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/tv_headcount"
-        tools:text="26,000원 (현장 결제)"
-        />
+        tools:text="26,000원 (현장 결제)" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
